+++ src/schemes/fluid/shifting_techniques.jl
@@ -42,60 +42,402 @@ end
 
 # Additional term in the momentum equation due to the shifting technique
 @inline function dv_shifting(shifting, system, neighbor_system,
-                             particle, neighbor, v_system, v_neighbor_system,
+                             v_system, v_neighbor_system, particle, neighbor,
                              m_a, m_b, rho_a, rho_b, pos_diff, distance,
                              grad_kernel, correction)
     return zero(grad_kernel)
 end
 
-@doc raw"""
-    ParticleShiftingTechnique()
+# Additional term(s) in the continuity equation due to the shifting technique
+function continuity_equation_shifting!(dv, shifting,
+                                       particle_system, neighbor_system,
+                                       particle, neighbor, grad_kernel, rho_a, rho_b, m_b)
+    return dv
+end
 
-Particle Shifting Technique by [Sun et al. (2017)](@cite Sun2017).
-Following the original paper, the callback is applied in every time step and not
-in every stage of a multi-stage time integration method to reduce the computational
-cost and improve the stability of the scheme.
+@doc raw"""
+    ParticleShiftingTechnique(; integrate_shifting_velocity=true,
+                              update_everystage=false,
+                              modify_continuity_equation=true,
+                              second_continuity_equation_term=ContinuityEquationTermSun2019(),
+                              momentum_equation_term=MomentumEquationTermSun2019(),
+                              v_max_factor=1, sound_speed_factor=0)
+
+Particle Shifting Technique by [Sun et al. (2017)](@cite Sun2017)
+and [Sun et al. (2019)](@cite Sun2019).
+The keyword arguments allow to choose between the original methods from the two papers
+and variants in between.
+
+The default values of the keyword arguments provide the version of shifting
+that we recommend based on our experiments.
+The default values are subject to change in future releases.
 
 See [Particle Shifting Technique](@ref shifting) for more information on the method.
+
+We provide the following convenience constructors for common variants of the method:
+- [`ParticleShiftingTechniqueSun2017()`](@ref):
+    Particle Shifting Technique by [Sun et al. (2017)](@cite Sun2017).
+    Shifting is applied as a position correction in a callback after each time step.
+    No additional terms are added to the momentum or continuity equations.
+- [`ConsistentShiftingSun2019()`](@ref):
+    Consistent Particle Shifting Technique by [Sun et al. (2019)](@cite Sun2019).
+    Shifting is applied with a shifting velocity in each stage of the time integration.
+    Additional terms are added to the momentum and continuity equations, most importantly
+    to guarantee conservation of volume in closed systems, which is not the case for the
+    original method by [Sun et al. (2017)](@cite Sun2017).
+
+# Keywords
+- `integrate_shifting_velocity`: If `true`, the shifting is applied in each stage of the
+                                time integration method as a shifting velocity that is
+                                added to the physical velocity in the time integration.
+                                If `false`, the shifting is applied as a position correction
+                                in a callback after each time step.
+- `update_everystage`:          If `true`, the shifting velocity is updated in every stage
+                                of a multi-stage time integration method.
+                                This requires `integrate_shifting_velocity=true`.
+                                If `false`, the shifting velocity is only updated once
+                                per time step in a callback, and the same shifting velocity
+                                is used for all stages.
+                                `update_everystage=false` reduces the computational cost,
+                                but may reduce the stability of the scheme and require
+                                a smaller time step.
+- `modify_continuity_equation`: If `true`, the continuity equation is modified to be based
+                                on the transport velocity instead of the physical velocity.
+                                This guarantees conservation of volume in closed systems,
+                                but is unstable at solid wall boundaries, according to our
+                                experiments.
+                                This requires `integrate_shifting_velocity=true`.
+- `second_continuity_equation_term`: Additional term to be added to the
+                                continuity equation to solve the stability problems with
+                                the modified continuity equation at solid wall boundaries.
+                                This requires `modify_continuity_equation=true`.
+                                See [`ContinuityEquationTermSun2019`](@ref).
+- `momentum_equation_term`:     Additional term to be added to the momentum equation
+                                to account for the shifting velocity.
+                                This requires `integrate_shifting_velocity=true`.
+                                See [`MomentumEquationTermSun2019`](@ref).
+- `v_max_factor`:               Factor to scale the expected maximum velocity used in the
+                                shifting velocity. The maximum expected velocity is computed as
+                                `v_max_factor * max(|v|)`, where `v` is the physical velocity.
+                                As opposed to `sound_speed_factor`, the computed expected
+                                maximum velocity depends on the current flow field
+                                and can change over time.
+                                Only one of `v_max_factor` and `sound_speed_factor`
+                                can be non-zero.
+- `sound_speed_factor`:         Factor to compute the maximum expected velocity used in the
+                                shifting velocity from the speed of sound.
+                                The maximum expected velocity is computed as
+                                `sound_speed_factor * c`, where `c` is the speed of sound.
+                                Only one of `v_max_factor` and `sound_speed_factor`
+                                can be non-zero.
 
 !!! warning
     The Particle Shifting Technique needs to be disabled close to the free surface
     and therefore requires a free surface detection method. This is not yet implemented.
     **This technique cannot be used in a free surface simulation.**
 """
-struct ParticleShiftingTechnique <: AbstractShiftingTechnique end
+struct ParticleShiftingTechnique{integrate_shifting_velocity,
+                                 update_everystage,
+                                 modify_continuity_equation,
+                                 compute_v_max,
+                                 ELTYPE, S, M} <: AbstractShiftingTechnique
+    v_factor                        :: ELTYPE
+    second_continuity_equation_term :: S
+    momentum_equation_term          :: M
 
-# Zero because PST is applied in a callback
-@inline function delta_v(system, ::ParticleShiftingTechnique, particle)
-    return zero(SVector{ndims(system), eltype(system)})
-end
+    function ParticleShiftingTechnique(; integrate_shifting_velocity=true,
+                                       update_everystage=false,
+                                       modify_continuity_equation=true,
+                                       second_continuity_equation_term=ContinuityEquationTermSun2019(),
+                                       momentum_equation_term=MomentumEquationTermSun2019(),
+                                       v_max_factor=1, sound_speed_factor=0)
+        if !integrate_shifting_velocity && update_everystage
+            throw(ArgumentError("ParticleShiftingTechnique: " *
+                                "integrate_shifting_velocity=false requires " *
+                                "update_everystage=false"))
+        end
 
-function particle_shifting_from_callback!(u_ode, shifting::ParticleShiftingTechnique,
-                                          system, v_ode, semi, dt)
-    @trixi_timeit timer() "particle shifting" begin
-        v = wrap_v(v_ode, system, semi)
-        u = wrap_u(u_ode, system, semi)
+        if !integrate_shifting_velocity && modify_continuity_equation
+            throw(ArgumentError("ParticleShiftingTechnique: " *
+                                "modify_continuity_equation=true requires " *
+                                "integrate_shifting_velocity=true"))
+        end
 
-        # Update the shifting velocity
-        update_shifting_from_callback!(system, shifting, v, u, v_ode, u_ode, semi)
+        if !modify_continuity_equation && !isnothing(second_continuity_equation_term)
+            throw(ArgumentError("ParticleShiftingTechnique: " *
+                                "a second_continuity_equation_term requires " *
+                                "modify_continuity_equation=true"))
+        end
 
-        # Update the particle positions with the shifting velocity
-        particle_shifting!(u_ode, shifting, system, semi, dt)
+        if !integrate_shifting_velocity && !isnothing(momentum_equation_term)
+            throw(ArgumentError("ParticleShiftingTechnique: " *
+                                "a momentum_equation_term requires " *
+                                "integrate_shifting_velocity=true"))
+        end
+
+        if v_max_factor > 0 && sound_speed_factor > 0
+            throw(ArgumentError("ParticleShiftingTechnique: " *
+                                "Only one of v_max_factor and sound_speed_factor " *
+                                "can be non-zero"))
+        end
+
+        if v_max_factor <= 0 && sound_speed_factor <= 0
+            throw(ArgumentError("ParticleShiftingTechnique: " *
+                                "One of v_max_factor and sound_speed_factor " *
+                                "must be positive"))
+        end
+
+        v_factor = max(v_max_factor, sound_speed_factor)
+        compute_v_max = v_max_factor > 0
+
+        new{integrate_shifting_velocity,
+            update_everystage,
+            modify_continuity_equation,
+            compute_v_max, typeof(v_factor),
+            typeof(second_continuity_equation_term),
+            typeof(momentum_equation_term)}(v_factor,
+                                            second_continuity_equation_term,
+                                            momentum_equation_term)
     end
 end
 
-function update_shifting_from_callback!(system, ::ParticleShiftingTechnique,
-                                        v, u, v_ode, u_ode, semi)
-    (; cache) = system
-    (; delta_v) = cache
+"""
+    ParticleShiftingTechniqueSun2017(; kwargs...)
 
-    set_zero!(delta_v)
+Particle Shifting Technique by [Sun et al. (2017)](@cite Sun2017).
+Following the original paper, the callback is applied in every time step and not
+in every stage of a multi-stage time integration method to reduce the computational cost.
 
+This is a convenience constructor for:
+```jldoctest; output = false
+ParticleShiftingTechnique(integrate_shifting_velocity=false,
+                          update_everystage=false,
+                          modify_continuity_equation=false,
+                          second_continuity_equation_term=nothing,
+                          momentum_equation_term=nothing,
+                          v_max_factor=1, sound_speed_factor=0)
+
+# output
+ParticleShiftingTechnique{false, false, false, true, Int64, Nothing, Nothing}(1, nothing, nothing)
+```
+
+See [ParticleShiftingTechnique](@ref ParticleShiftingTechnique) for all available options.
+
+# Keywords
+- `kwargs...`: All keywords are passed to the main constructor.
+
+# Examples
+```jldoctest; output = false
+shifting_technique = ParticleShiftingTechniqueSun2017()
+
+# output
+ParticleShiftingTechnique{false, false, false, true, Int64, Nothing, Nothing}(1, nothing, nothing)
+```
+
+!!! warning
+    The Particle Shifting Technique needs to be disabled close to the free surface
+    and therefore requires a free surface detection method. This is not yet implemented.
+    **This technique cannot be used in a free surface simulation.**
+"""
+function ParticleShiftingTechniqueSun2017(; kwargs...)
+    return ParticleShiftingTechnique(; integrate_shifting_velocity=false,
+                                     update_everystage=false,
+                                     modify_continuity_equation=false,
+                                     second_continuity_equation_term=nothing,
+                                     momentum_equation_term=nothing,
+                                     v_max_factor=1, sound_speed_factor=0,
+                                     kwargs...)
+end
+
+"""
+    ConsistentShiftingSun2019(; sound_speed_factor=0.1, kwargs...)
+
+Consistent Particle Shifting Technique by [Sun et al. (2019)](@cite Sun2019).
+
+This is a convenience constructor for:
+```jldoctest; output = false
+ParticleShiftingTechnique(integrate_shifting_velocity=true,
+                          update_everystage=true,
+                          modify_continuity_equation=true,
+                          second_continuity_equation_term=ContinuityEquationTermSun2019(),
+                          momentum_equation_term=MomentumEquationTermSun2019(),
+                          v_max_factor=0, sound_speed_factor=0.1)
+
+# output
+ParticleShiftingTechnique{true, true, true, false, Float64, ContinuityEquationTermSun2019, MomentumEquationTermSun2019}(0.1, ContinuityEquationTermSun2019(), MomentumEquationTermSun2019())
+```
+
+See [ParticleShiftingTechnique](@ref ParticleShiftingTechnique) for all available options.
+
+# Keywords
+- `sound_speed_factor`: Factor to compute the maximum expected velocity used in the
+                        shifting velocity from the speed of sound.
+                        The maximum expected velocity is computed as
+                        `sound_speed_factor * c`, where `c` is the speed of sound.
+                        Since the speed of sound is usually chosen as 10 times the maximum
+                        expected velocity in weakly compressible SPH, a value of 0.1
+                        corresponds to the maximum expected velocity.
+- `kwargs...`: All keywords are passed to the main constructor.
+
+# Examples
+```jldoctest; output = false
+shifting_technique = ConsistentShiftingSun2019()
+
+# output
+ParticleShiftingTechnique{true, true, true, false, Float64, ContinuityEquationTermSun2019, MomentumEquationTermSun2019}(0.1, ContinuityEquationTermSun2019(), MomentumEquationTermSun2019())
+```
+
+!!! warning
+    The Particle Shifting Technique needs to be disabled close to the free surface
+    and therefore requires a free surface detection method. This is not yet implemented.
+    **This technique cannot be used in a free surface simulation.**
+"""
+function ConsistentShiftingSun2019(; kwargs...)
+    return ParticleShiftingTechnique(; integrate_shifting_velocity=true,
+                                     update_everystage=true,
+                                     modify_continuity_equation=true,
+                                     second_continuity_equation_term=ContinuityEquationTermSun2019(),
+                                     momentum_equation_term=MomentumEquationTermSun2019(),
+                                     v_max_factor=0, sound_speed_factor=0.1,
+                                     kwargs...)
+end
+
+# `ParticleShiftingTechnique{false}` means `integrate_shifting_velocity=false`.
+# Zero if PST is applied in a callback as a position correction
+# and not with a shifting velocity in the time integration stages
+# (which would be `integrate_shifting_velocity=false`).
+@inline function delta_v(system, ::ParticleShiftingTechnique{false}, particle)
+    return zero(SVector{ndims(system), eltype(system)})
+end
+
+"""
+    MomentumEquationTermSun2019()
+
+A term by [Sun et al. (2019)](@cite Sun2019) to be added to the momentum equation.
+
+See [`ParticleShiftingTechnique`](@ref).
+"""
+struct MomentumEquationTermSun2019 end
+
+# Additional term in the momentum equation due to the shifting technique
+@inline function dv_shifting(shifting::ParticleShiftingTechnique, system, neighbor_system,
+                             v_system, v_neighbor_system, particle, neighbor,
+                             m_a, m_b, rho_a, rho_b, pos_diff, distance,
+                             grad_kernel, correction)
+    return dv_shifting(shifting.momentum_equation_term, system, neighbor_system,
+                       v_system, v_neighbor_system, particle, neighbor,
+                       m_a, m_b, rho_a, rho_b, pos_diff, distance,
+                       grad_kernel, correction)
+end
+
+@propagate_inbounds function dv_shifting(::MomentumEquationTermSun2019,
+                                         system, neighbor_system,
+                                         v_system, v_neighbor_system,
+                                         particle, neighbor, m_a, m_b, rho_a, rho_b,
+                                         pos_diff, distance, grad_kernel, correction)
+    delta_v_a = delta_v(system, particle)
+    delta_v_b = delta_v(neighbor_system, neighbor)
+
+    v_a = current_velocity(v_system, system, particle)
+    v_b = current_velocity(v_neighbor_system, neighbor_system, neighbor)
+
+    tensor_product = v_a * delta_v_a' + v_b * delta_v_b'
+    return m_b / rho_b *
+           (tensor_product * grad_kernel + v_a * dot(delta_v_a - delta_v_b, grad_kernel))
+end
+
+# `ParticleShiftingTechnique{<:Any, <:Any, true}` means `modify_continuity_equation=true`
+function continuity_equation_shifting!(dv,
+                                       shifting::ParticleShiftingTechnique{<:Any, <:Any,
+                                                                           true},
+                                       system, neighbor_system,
+                                       particle, neighbor, grad_kernel, rho_a, rho_b, m_b)
+    delta_v_diff = delta_v(system, particle) -
+                   delta_v(neighbor_system, neighbor)
+
+    dv[end, particle] += rho_a / rho_b * m_b * dot(delta_v_diff, grad_kernel)
+
+    second_continuity_equation_term!(dv, shifting.second_continuity_equation_term,
+                                     system, neighbor_system,
+                                     particle, neighbor, grad_kernel, rho_a, rho_b, m_b)
+
+    return dv
+end
+
+"""
+    ContinuityEquationTermSun2019()
+
+A second term by [Sun et al. (2019)](@cite Sun2019) to be added to the continuity equation
+to solve the stability problems with the modified continuity equation at solid wall boundaries.
+
+See [`ParticleShiftingTechnique`](@ref).
+"""
+struct ContinuityEquationTermSun2019 end
+
+@inline function second_continuity_equation_term!(dv,
+                                                  ::ContinuityEquationTermSun2019,
+                                                  system, neighbor_system,
+                                                  particle, neighbor, grad_kernel,
+                                                  rho_a, rho_b, m_b)
+    rho_v = rho_a * delta_v(system, particle) + rho_b * delta_v(neighbor_system, neighbor)
+
+    dv[end, particle] += m_b / rho_b * dot(rho_v, grad_kernel)
+
+    return dv
+end
+
+@inline function second_continuity_equation_term!(dv, second_continuity_equation_term,
+                                                  system, neighbor_system,
+                                                  particle, neighbor, grad_kernel,
+                                                  rho_a, rho_b, m_b)
+    return dv
+end
+
+# `ParticleShiftingTechnique{<:Any, true}` means `update_everystage=true`
+function update_shifting!(system, shifting::ParticleShiftingTechnique{<:Any, true},
+                          v, u, v_ode, u_ode, semi)
+    update_shifting_inner!(system, shifting, v, u, v_ode, u_ode, semi)
+end
+
+# `ParticleShiftingTechnique{<:Any, false}` means `update_everystage=false`
+function update_shifting_from_callback!(system,
+                                        shifting::ParticleShiftingTechnique{<:Any, false},
+                                        v_ode, u_ode, semi)
+    v = wrap_v(v_ode, system, semi)
+    u = wrap_u(u_ode, system, semi)
+
+    update_shifting_inner!(system, shifting, v, u, v_ode, u_ode, semi)
+end
+
+# `ParticleShiftingTechnique{<:Any, <:Any, <:Any, true}`
+# means `compute_v_max=true`
+function v_max(shifting::ParticleShiftingTechnique{<:Any, <:Any, <:Any, true},
+               v, system)
     # This has similar performance to `maximum(..., eachparticle(system))`,
     # but is GPU-compatible.
     v_max = maximum(x -> sqrt(dot(x, x)),
                     reinterpret(reshape, SVector{ndims(system), eltype(v)},
                                 current_velocity(v, system)))
+    return shifting.v_factor * v_max
+end
+
+# `ParticleShiftingTechnique{<:Any, <:Any, <:Any, false}`
+# means `compute_v_max=false`
+function v_max(shifting::ParticleShiftingTechnique{<:Any, <:Any, <:Any, false},
+               v, system)
+    sound_speed = system_sound_speed(system)
+
+    return shifting.v_factor * sound_speed
+end
+
+function update_shifting_inner!(system, shifting::ParticleShiftingTechnique,
+                                v, u, v_ode, u_ode, semi)
+    (; cache) = system
+    (; delta_v) = cache
+
+    set_zero!(delta_v)
+
+    v_max_ = v_max(shifting, v, system)
 
     # TODO this needs to be adapted to multi-resolution.
     # Section 3.2 explains what else needs to be changed.
@@ -127,7 +469,8 @@ function update_shifting_from_callback!(system, ::ParticleShiftingTechnique,
 
             # Eq. 7 in Sun et al. (2017).
             # According to the paper, CFL * Ma can be rewritten as Δt * v_max / h
-            # (see p. 29, right above Eq. 9), but this does not work when scaling h.
+            # (see p. 29, right above Eq. 9), but this does not yield the same amount
+            # of shifting when scaling h.
             # When setting CFL * Ma = Δt * v_max / (2 * Δx), PST works as expected
             # for both small and large smoothing length factors.
             # We need to scale
@@ -135,7 +478,7 @@ function update_shifting_from_callback!(system, ::ParticleShiftingTechnique,
             # - linearly with the particle spacing,
             # - linearly with the time step.
             # See https://github.com/trixi-framework/TrixiParticles.jl/pull/834.
-            delta_v_ = -v_max * (2 * h)^2 / (2 * dx) * (1 + R * (kernel / Wdx)^n) *
+            delta_v_ = -v_max_ * (2 * h)^2 / (2 * dx) * (1 + R * (kernel / Wdx)^n) *
                        m_b / (rho_a + rho_b) * grad_kernel
 
             # Write into the buffer
@@ -148,7 +491,28 @@ function update_shifting_from_callback!(system, ::ParticleShiftingTechnique,
     return system
 end
 
-function particle_shifting!(u_ode, ::ParticleShiftingTechnique, system, semi, dt)
+# `ParticleShiftingTechnique{<:Any, false}` means `update_everystage=false`.
+# Only update shifting from callback if `update_everystage=false`.
+# Only apply shifting from callback if PST is to be applied in a callback
+# (`integrate_shifting_velocity=false`), but this also requires `update_everystage=false`.
+function particle_shifting_from_callback!(u_ode,
+                                          shifting::ParticleShiftingTechnique{<:Any, false},
+                                          system, v_ode, semi, dt)
+    @trixi_timeit timer() "particle shifting" begin
+        # Update the shifting velocity
+        update_shifting_from_callback!(system, shifting, v_ode, u_ode, semi)
+
+        # Update the particle positions with the shifting velocity
+        apply_particle_shifting!(u_ode, shifting, system, semi, dt)
+    end
+end
+
+# `ParticleShiftingTechnique{false}` means `integrate_shifting_velocity=false`.
+# Only apply shifting from callback if PST is to be applied in a callback
+# and not with a shifting velocity in the time integration stages
+# (which would be `integrate_shifting_velocity=false`).
+function apply_particle_shifting!(u_ode, ::ParticleShiftingTechnique{false},
+                                  system, semi, dt)
     (; cache) = system
     (; delta_v) = cache
 
@@ -164,30 +528,45 @@ function particle_shifting!(u_ode, ::ParticleShiftingTechnique, system, semi, dt
     return u
 end
 
+function apply_particle_shifting!(u_ode, ::ParticleShiftingTechnique{true},
+                                  system, semi, dt)
+    return u_ode
+end
+
 """
-    TransportVelocityAdami(background_pressure::Real)
+    TransportVelocityAdami(; background_pressure::Real, modify_continuity_equation=false)
 
 Transport Velocity Formulation (TVF) by [Adami et al. (2013)](@cite Adami2013)
 to suppress pairing and tensile instability.
 See [TVF](@ref transport_velocity_formulation) for more details of the method.
 
-# Arguments
+# Keywords
 - `background_pressure`: Background pressure. Suggested is a background pressure which is
                          on the order of the reference pressure.
+- `modify_continuity_equation`: If `true`, the continuity equation is modified to be based
+                                on the transport velocity instead of the physical velocity.
+                                This guarantees conservation of volume in closed systems,
+                                but is unstable at solid wall boundaries, according to our
+                                experiments.
 
 !!! warning
     The Transport Velocity Formulation needs to be disabled close to the free surface
     and therefore requires a free surface detection method. This is not yet implemented.
     **This technique cannot be used in a free surface simulation.**
 """
-struct TransportVelocityAdami{T <: Real} <: AbstractShiftingTechnique
+struct TransportVelocityAdami{modify_continuity_equation, T <: Real} <:
+       AbstractShiftingTechnique
     background_pressure::T
+
+    function TransportVelocityAdami(; background_pressure, modify_continuity_equation=false)
+        new{modify_continuity_equation, typeof(background_pressure)}(background_pressure)
+    end
 end
 
-@inline function dv_shifting(::TransportVelocityAdami, system, neighbor_system,
-                             particle, neighbor, v_system, v_neighbor_system,
-                             m_a, m_b, rho_a, rho_b, pos_diff, distance,
-                             grad_kernel, correction)
+@propagate_inbounds function dv_shifting(::TransportVelocityAdami, system, neighbor_system,
+                                         v_system, v_neighbor_system, particle, neighbor,
+                                         m_a, m_b, rho_a, rho_b, pos_diff, distance,
+                                         grad_kernel, correction)
     v_a = current_velocity(v_system, system, particle)
     delta_v_a = delta_v(system, particle)
 
@@ -207,6 +586,17 @@ end
     return pressure_acceleration(system, neighbor_system, particle, neighbor,
                                  m_a, m_b, A_a, A_b, rho_a, rho_b, pos_diff,
                                  distance, grad_kernel, correction)
+end
+
+function continuity_equation_shifting!(dv, shifting::TransportVelocityAdami{true},
+                                       particle_system, neighbor_system,
+                                       particle, neighbor, grad_kernel, rho_a, rho_b, m_b)
+    delta_v_diff = delta_v(particle_system, particle) -
+                   delta_v(neighbor_system, neighbor)
+
+    dv[end, particle] += rho_a / rho_b * m_b * dot(delta_v_diff, grad_kernel)
+
+    return dv
 end
 
 function update_shifting!(system, shifting::TransportVelocityAdami, v, u, v_ode,